+++ tom_github_api/lib/tom_github_api.dart
@@ -17,6 +17,7 @@ export 'src/models/github_comment.dart';
 export 'src/models/github_conditional.dart';
 export 'src/models/github_content.dart';
 export 'src/models/github_git_objects.dart';
+export 'src/models/github_graphql.dart';
 export 'src/models/github_issue.dart';
 export 'src/models/github_label.dart';
 export 'src/models/github_rate_limit.dart';